+++ src/test/java/io/supertokens/test/jwt/api/JWTSigningAPITest2_19.java
@@ -1,0 +1,127 @@
+/*
+ *    Copyright (c) 2021, VRAI Labs and/or its affiliates. All rights reserved.
+ *
+ *    This software is licensed under the Apache License, Version 2.0 (the
+ *    "License") as published by the Apache Software Foundation.
+ *
+ *    You may not use this file except in compliance with the License. You may
+ *    obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ *    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ *    License for the specific language governing permissions and limitations
+ *    under the License.
+ */
+
+package io.supertokens.test.jwt.api;
+
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.interfaces.Claim;
+import com.auth0.jwt.interfaces.DecodedJWT;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonNull;
+import com.google.gson.JsonObject;
+import io.supertokens.ProcessState;
+import io.supertokens.test.TestingProcessManager;
+import io.supertokens.test.Utils;
+import io.supertokens.test.httpRequest.HttpRequestForTesting;
+import io.supertokens.test.httpRequest.HttpResponseException;
+import io.supertokens.utils.SemVer;
+import io.supertokens.webserver.api.jwt.JWTSigningAPI;
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TestRule;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.junit.Assert.*;
+
+public class JWTSigningAPITest2_19 {
+    @Rule
+    public TestRule watchman = Utils.getOnFailure();
+
+    @AfterClass
+    public static void afterTesting() {
+        Utils.afterTesting();
+    }
+
+    @Before
+    public void beforeEach() {
+        Utils.reset();
+    }
+
+    @Test
+    public void testThatStaticSignedJWTHasValidHeader() throws Exception {
+        String[] args = { "../" };
+
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        JsonObject requestBody = new JsonObject();
+        requestBody.addProperty("algorithm", "rs256");
+        requestBody.addProperty("jwksDomain", "http://localhost");
+        requestBody.addProperty("useStaticSigningKey", true);
+        requestBody.add("payload", new JsonObject());
+        requestBody.addProperty("validity", 3600);
+
+        JsonObject response = HttpRequestForTesting.sendJsonPOSTRequest(process.getProcess(), "",
+                "http://localhost:3567/recipe/jwt", requestBody, 1000, 1000, null, SemVer.v2_19.get(),
+                "jwt");
+
+        String jwt = response.get("jwt").getAsString();
+        DecodedJWT decodedJWT = JWT.decode(jwt);
+
+        Claim headerAlg = decodedJWT.getHeaderClaim("alg");
+        Claim headerType = decodedJWT.getHeaderClaim("typ");
+        Claim headerKeyId = decodedJWT.getHeaderClaim("kid");
+
+        assertTrue(!headerAlg.isNull() && !headerType.isNull() && !headerKeyId.isNull());
+        assert headerAlg.asString().equalsIgnoreCase("rs256");
+        assert headerType.asString().equalsIgnoreCase("jwt");
+        assert !headerKeyId.asString().isEmpty();
+        assert headerKeyId.asString().startsWith("s-");
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testThatDynamicSignedJWTHasValidHeader() throws Exception {
+        String[] args = { "../" };
+
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        JsonObject requestBody = new JsonObject();
+        requestBody.addProperty("algorithm", "rs256");
+        requestBody.addProperty("jwksDomain", "http://localhost");
+        requestBody.addProperty("useStaticSigningKey", false);
+        requestBody.add("payload", new JsonObject());
+        requestBody.addProperty("validity", 3600);
+
+        JsonObject response = HttpRequestForTesting.sendJsonPOSTRequest(process.getProcess(), "",
+                "http://localhost:3567/recipe/jwt", requestBody, 1000, 1000, null, SemVer.v2_19.get(),
+                "jwt");
+
+        String jwt = response.get("jwt").getAsString();
+        DecodedJWT decodedJWT = JWT.decode(jwt);
+
+        Claim headerAlg = decodedJWT.getHeaderClaim("alg");
+        Claim headerType = decodedJWT.getHeaderClaim("typ");
+        Claim headerKeyId = decodedJWT.getHeaderClaim("kid");
+
+        assertTrue(!headerAlg.isNull() && !headerType.isNull() && !headerKeyId.isNull());
+        assert headerAlg.asString().equalsIgnoreCase("rs256");
+        assert headerType.asString().equalsIgnoreCase("jwt");
+        assert !headerKeyId.asString().isEmpty();
+        assert headerKeyId.asString().startsWith("d-");
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+}