+++ src/test/java/io/supertokens/test/session/AccessTokenSigningKeyTest.java
@@ -75,20 +75,18 @@ public class AccessTokenSigningKeyTest {
         sessionStorage.setKeyValue("access_token_signing_key", newKey);
         AccessTokenSigningKey accessTokenSigningKeyInstance = AccessTokenSigningKey.getInstance(process.getProcess());
         accessTokenSigningKeyInstance.transferLegacyKeyToNewTable();
-        assertEquals(SigningKeys.getInstance(process.getProcess()).getAllKeys().size(), 3);
-        List<JWTSigningKeyInfo> keys = SigningKeys.getInstance(process.getProcess()).getStaticKeys();
-        assertEquals(keys.size(), 2);
+        assertEquals(SigningKeys.getInstance(process.getProcess()).getAllKeys().size(), 2);
+        List<KeyInfo> keys = SigningKeys.getInstance(process.getProcess()).getDynamicKeys();
+        assertEquals(keys.size(), 1);
         assertEquals(keys.get(0).createdAtTime, newKey.createdAtTime);
-        assertEquals(keys.get(0).keyString, newKey.value);
+        assertEquals(keys.get(0).value, newKey.value);
         assertEquals(sessionStorage.getKeyValue("access_token_signing_key"), null);
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(PROCESS_STATE.STOPPED));
     }
 
     @Test
-    public void getDynamicKeysReturnsOrdered()
-            throws IOException, InterruptedException, InvalidKeyException, NoSuchAlgorithmException,
-            StorageQueryException, StorageTransactionLogicException, InvalidKeySpecException, SignatureException {
+    public void getDynamicKeysReturnsOrdered() throws Exception {
         Utils.setValueInConfig("access_token_dynamic_signing_key_update_interval", "0.00027"); // 1 seconds
 
         String[] args = { "../" };
@@ -99,8 +97,8 @@ public class AccessTokenSigningKeyTest {
 
         io.supertokens.utils.Utils.PubPriKey rsaKeys = io.supertokens.utils.Utils.generateNewPubPriKey();
         String signingKey = rsaKeys.toString();
-        KeyValueInfo legacyKey = new KeyValueInfo(signingKey, System.currentTimeMillis() - 2000); // 2 seconds in the
-                                                                                                  // past
+        KeyValueInfo legacyKey = new KeyValueInfo(signingKey, System.currentTimeMillis() - 2000);
+        // 2 seconds in the past
 
         SessionStorage sessionStorage = StorageLayer.getSessionStorage(process.getProcess());
         sessionStorage.setKeyValue("access_token_signing_key", legacyKey);
@@ -113,23 +111,22 @@ public class AccessTokenSigningKeyTest {
         // Wait for access_token_dynamic_signing_key_update_interval + margin
         Thread.sleep(1500);
 
-        List<KeyInfo> allKeys = SigningKeys.getInstance(process.getProcess()).getDynamicKeys();
-        // We get a migrated + new static signing key (generated on startup)
-        // + 1 expired and 1 fresh dynamic key
-        assertEquals(allKeys.size(), 4);
+        List<KeyInfo> dynamicKeys = SigningKeys.getInstance(process.getProcess()).getDynamicKeys();
+        // We get a migrated + 1 expired and 1 fresh dynamic key
+        assertEquals(dynamicKeys.size(), 3);
 
         // The first one should be the latest key
         KeyInfo key = SigningKeys.getInstance(process.getProcess()).getLatestIssuedDynamicKey();
-        assertEquals(allKeys.get(0).createdAtTime, key.createdAtTime);
-        assertEquals(allKeys.get(0).value, key.value);
+        assertEquals(dynamicKeys.get(0).createdAtTime, key.createdAtTime);
+        assertEquals(dynamicKeys.get(0).value, key.value);
 
         // The oldest one should be the legacy key.
-        assertEquals(allKeys.get(3).createdAtTime, legacyKey.createdAtTime);
-        assertEquals(allKeys.get(3).value, legacyKey.value);
+        assertEquals(dynamicKeys.get(2).createdAtTime, legacyKey.createdAtTime);
+        assertEquals(dynamicKeys.get(2).value, legacyKey.value);
 
         // Keys should be ordered by createdAtTime descending
-        for (int i = 0; i < allKeys.size() - 1; ++i) {
-            assertTrue(allKeys.get(i).createdAtTime > allKeys.get(i + 1).createdAtTime);
+        for (int i = 0; i < dynamicKeys.size() - 1; ++i) {
+            assertTrue(dynamicKeys.get(i).createdAtTime > dynamicKeys.get(i + 1).createdAtTime);
         }
 
         process.kill();
@@ -193,7 +190,7 @@ public class AccessTokenSigningKeyTest {
         assertEquals(keys.size(), 2);
 
         assertEquals(legacyKey.value, keys.get(1).keyString);
-
+        assertEquals(keys.get(1).keyId.substring(0, 2), "s-");
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(PROCESS_STATE.STOPPED));
     }